Update: New main file
This represents the new entry file. It's been changed to
accept paired sequencing fasta files as well. Set --paired
flagged if the mapping includes paired reads.

Moreover, the filter and the count procedure has been updated
to account for the whole genome. A bed file is therefore not
required. This has been indeed chosen to be the identifier for
(a) running the slamdunk pipeline wrt some reference positions
or (b) running it wrt the whole genome. If you don't pass a bed
file, the latter is chosen.