Bugfix: Sort file before iterating
The file is sorted and stored in the correct folder
before the counting takes place. Sorting by the name is necessary
to facilitate the paired read iteration. If there's already a file
w/ a _sorted suffix, the file is loaded instead. If file is already
sorted, no additional sorting is performed